requirements.txt
torch>=1.13.0
torchvision>=0.14.0
torchaudio>=0.13.0
timm==0.4.12
mamba_ssm==1.0.1
causal_conv1d==1.0.0
tensorboard
scikit-learn
matplotlib
pandas
Pillow
tqdm
packaging
triton==2.0.0
thop
h5py
SimpleITK
scikit-image
medpy
yacs
